Auto stash before merge of "hotfix/security_exception" and "origin/develop"
